+++ bolshakovPetr/app/(tabs)/_layout.tsx
@@ -1,37 +1,49 @@
-import { Tabs } from 'expo-router';
-import React from 'react';
-
-import { TabBarIcon } from '@/components/navigation/TabBarIcon';
-import { Colors } from '@/constants/Colors';
-import { useColorScheme } from '@/hooks/useColorScheme';
+import { Tabs } from "expo-router";
+import React from "react";
+import "react-native-reanimated";
+import { TabBarIcon } from "@/components/navigation/TabBarIcon";
+import { Colors } from "@/constants/Colors";
+import { useColorScheme } from "@/hooks/useColorScheme";
+import Entypo from '@expo/vector-icons/Entypo';
 
 export default function TabLayout() {
   const colorScheme = useColorScheme();
 
   return (
-    <Tabs
-      screenOptions={{
-        tabBarActiveTintColor: Colors[colorScheme ?? 'light'].tint,
-        headerShown: false,
-      }}>
-      <Tabs.Screen
-        name="index"
-        options={{
-          title: 'Home',
-          tabBarIcon: ({ color, focused }) => (
-            <TabBarIcon name={focused ? 'home' : 'home-outline'} color={color} />
-          ),
+    <>
+      <Tabs
+        screenOptions={{
+          tabBarActiveTintColor: Colors[colorScheme ?? "light"].tint,
         }}
-      />
-      <Tabs.Screen
-        name="explore"
-        options={{
-          title: 'Explore',
-          tabBarIcon: ({ color, focused }) => (
-            <TabBarIcon name={focused ? 'code-slash' : 'code-slash-outline'} color={color} />
-          ),
-        }}
-      />
-    </Tabs>
+      >
+        <Tabs.Screen
+          name="lab1"
+          options={{
+            title: "Lab1",
+            tabBarIcon: ({ color, size }) => (
+              <Entypo name="lab-flask" size={size} color={color}/>
+            ),
+          }}
+        />
+        <Tabs.Screen
+          name="lab2"
+          options={{
+            title: "Lab2",
+            tabBarIcon: ({ color, size }) => (
+              <Entypo name="lab-flask" size={size} color={color}/>
+            ),
+          }}
+        />
+        <Tabs.Screen
+          name="lab3"
+          options={{
+            title: "Lab3",
+            tabBarIcon: ({ color, size }) => (
+              <Entypo name="lab-flask" size={size} color={color}/>
+            ),
+          }}
+        />
+      </Tabs>
+    </>
   );
 }
