+++ bolshakovPetr/app/(tabs)/lab3.tsx
@@ -1,0 +1,31 @@
+import {
+  Image,
+  StyleSheet,
+  Platform,
+  SafeAreaView,
+  View,
+  Text,
+  TouchableOpacity,
+} from "react-native";
+import React, { useEffect, useState } from "react";
+import { HelloWave } from "@/components/HelloWave";
+import ParallaxScrollView from "@/components/ParallaxScrollView";
+import { ThemedText } from "@/components/ThemedText";
+import { ThemedView } from "@/components/ThemedView";
+import axios from "axios";
+
+export default function Lab3() {
+  return (
+    <SafeAreaView style={{ flex: 1, display: "flex" }}>
+      <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+        <View
+          style={{
+            backgroundColor: colors[colorIndex],
+            width: boxSize,
+            height: boxSize,
+          }}
+        />
+      </View>
+    </SafeAreaView>
+  );
+}
